ibarr/path_reslove: Update sm_lid and sm_sl on events
It's possible that ibarr is started before opensm is started, in this case
ibarr needs to update sm_lid as sm_sl based on events, otherwise the sm_lid
could be 0.

This patch also replace umad_get_port() with ibv_query_port() to get
sm_lid and sm_sl, which looks like simpler.

Signed-off-by: Mark Zhang <markzhang@nvidia.com>