set_packages_version: set versions from cmake ver
Allow a single command to set packages (rpm and deb) version.
Doing this in cmake is wrong, as it causes problems for building the
package from git:

you don't have to have a proper build environment with all the relevant
libraries just to run that update. The downside is that currently
versions will probably need to be included in the the version control,
which is not ideal.

Signed-off-by: Tzafrir Cohen <nvidia@cohens.org.il>